requirements.txt
transformers==4.35.0
torch==2.0.1
fastapi==0.104.1
uvicorn==0.23.2
streamlit==1.28.0
faiss-cpu==1.7.4
langchain==0.0.312
peft==0.5.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.1
sentence-transformers==2.2.2
python-dotenv==1.0.0